pyproject.toml
[tool.poetry]
name = "clips-metal-es"
version = "0.1.0"
description = ""
authors = ["Shah Azizie <shahazizie360@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
clipspy = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
